# specs/001-mvp-skill-maps/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SKILL MAPS API
  version: 1.0.0
  description: Contract for the SKILL MAPS MVP web application.
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: /api
security: []
...
paths:
...
  /v1/me/certification-records:
    get:
      tags: [Credentials]
      summary: List certification acquisitions and renewals
      operationId: listCertificationRecords
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          description: All self-declared acquisitions and renewals.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CertificationRecord' }
        '401': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Credentials]
      summary: Record a certification acquisition or renewal
      operationId: createCertificationRecord
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CertificationRecordInput' }
      responses:
        '201':
          description: Separate acquisition or renewal created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CertificationRecord' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
...
  /v1/admin/trails:
    post:
      tags: [Administration]
      summary: Create a draft learning trail
      operationId: createTrail
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TrailDraftInput' }
      responses:
        '201':
          description: Draft trail created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminCatalogResource' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
        '429': { $ref: '#/components/responses/Problem' }
  /v1/admin/trails/{trailId}:
    patch:
      tags: [Administration]
      summary: Update a draft learning trail
      operationId: updateTrailDraft
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TrailId'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/TrailDraftInput' }
      responses:
        '200':
          description: Draft trail updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminCatalogResource' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
        '429': { $ref: '#/components/responses/Problem' }
  /v1/admin/trails/{trailId}/publications:
    post:
      tags: [Administration]
      summary: Publish a complete immutable trail revision
      operationId: publishTrail
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TrailId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '201':
          description: Immutable revision published and made current.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Publication' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
        '429': { $ref: '#/components/responses/Problem' }
  /v1/admin/certifications:
    post:
      tags: [Administration]
      summary: Create a draft certification
      operationId: createCertification
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CertificationInput' }
      responses:
        '201':
          description: Draft certification created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminCatalogResource' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
        '429': { $ref: '#/components/responses/Problem' }
  /v1/admin/achievements:
    post:
      tags: [Administration]
      summary: Create a draft achievement
      operationId: createAchievement
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AchievementInput' }
      responses:
        '201':
          description: Draft achievement created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminCatalogResource' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/ValidationProblem' }
        '429': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-skillmaps-session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
...
    TrailId:
      name: trailId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: RFC 9457 problem response.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationProblem:
      description: Input failed validation.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        requestId: { type: string }
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              maxItems: 50
              items:
                type: object
                required: [path, code]
                properties:
                  path: { type: string }
                  code: { type: string }
                  message: { type: string }
...
    CertificationRecordInput:
      type: object
      additionalProperties: false
      required: [certificationId, obtainedOn]
      properties:
        certificationId: { type: string, format: uuid }
        obtainedOn: { type: string, format: date }
        externalIdentifier: { type: [string, 'null'], maxLength: 180 }
        expiresOn: { type: [string, 'null'], format: date }
    CertificationRecord:
      allOf:
        - $ref: '#/components/schemas/CertificationRecordInput'
        - type: object
          required: [id, verificationStatus, createdAt]
          properties:
            id: { type: string, format: uuid }
            verificationStatus: { const: self_declared }
            createdAt: { type: string, format: date-time }
...
    TrailDraftInput:
      type: object
      additionalProperties: false
      properties:
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        categoryId: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 180 }
        description: { type: string, minLength: 1, maxLength: 8000 }
        targetRoleIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
        steps:
          type: array
          maxItems: 200
          items:
            type: object
            required: [stepId, position, title, required, skillIds, prerequisiteStepIds]
            properties:
              stepId: { type: string, format: uuid }
              position: { type: integer, minimum: 1 }
              title: { type: string, minLength: 1, maxLength: 180 }
              description: { type: string, maxLength: 4000 }
              required: { type: boolean }
              skillIds:
                type: array
                uniqueItems: true
                items: { type: string, format: uuid }
              prerequisiteStepIds:
                type: array
                uniqueItems: true
                items: { type: string, format: uuid }
    CertificationInput:
      type: object
      additionalProperties: false
      required: [slug, name, issuer, description]
      properties:
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        name: { type: string, minLength: 1, maxLength: 180 }
        issuer: { type: string, minLength: 1, maxLength: 180 }
        description: { type: string, minLength: 1, maxLength: 8000 }
        defaultValidityMonths: { type: integer, minimum: 1, maximum: 1200 }
        skillIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
        trailIds:
          type: array
          uniqueItems: true
          items: { type: string, format: uuid }
        requirements:
          type: array
          maxItems: 200
          items: { $ref: '#/components/schemas/CertificationRequirementInput' }
    CertificationRequirementInput:
      type: object
      additionalProperties: false
      required: [title, type, targetId, required, position]
      properties:
        title: { type: string, minLength: 1, maxLength: 180 }
        type: { type: string, enum: [skill, trail] }
        targetId: { type: string, format: uuid }
        required: { type: boolean }
        position: { type: integer, minimum: 1, maximum: 200 }
    AchievementInput:
      type: object
      additionalProperties: false
      required: [slug, title, description, iconLabel, criterionType, criterionParameters]
      properties:
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        title: { type: string, minLength: 1, maxLength: 180 }
        description: { type: string, minLength: 1, maxLength: 2000 }
        iconLabel: { type: string, minLength: 1, maxLength: 120 }
        criterionType:
          type: string
          enum: [completed_steps, certification_records]
        criterionParameters: { $ref: '#/components/schemas/AchievementCriterionParameters' }
    AchievementCriterionParameters:
      type: object
      additionalProperties: false
      required: [minimum]
      properties:
        minimum: { type: integer, minimum: 1, maximum: 1000 }
...
    AdminCatalogResource:
      type: object
      required: [id, slug, status, updatedAt]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        status: { type: string, enum: [draft, published, unpublished, inactive] }
        updatedAt: { type: string, format: date-time }
    Publication:
      type: object
      required: [resourceId, revisionId, revisionNumber, publishedAt]
      properties:
        resourceId: { type: string, format: uuid }
        revisionId: { type: string, format: uuid }
        revisionNumber: { type: integer, minimum: 1 }
        publishedAt: { type: string, format: date-time }